+++ src/main/object/sm_object.h
@@ -8,6 +8,6 @@ typedef struct sm_object {
 sm_string *sm_object_to_string(sm_object *obj1);
 uint32_t   sm_object_sprint(sm_object *obj1, char *buffer, bool fake);
 int        sm_sizeof(sm_object *sm_obj);
-bool       sm_object_is_literal(uint16_t t);
+bool       sm_object_is_literal(uint32_t t);
 void       sm_object_dbg_print(sm_object *obj);
 bool       sm_object_eq(sm_object *result, sm_object *self);
